requirements.txt
boto3==1.28.3
botocore==1.31.3
python-dotenv==1.0.0
